メモを追加する処理をクラスに分ける
