+++ 04.class/memo.js
@@ -3,6 +3,7 @@ import { MemoApp } from "./lib/memo_app.js";
 import { ListMemosCommand } from "./lib/list_memos_command.js";
 import { ReadMemoCommand } from "./lib/read_memo_command.js";
 import { DeleteMemoCommand } from "./lib/delete_memo_command.js";
+import { InsertMemoCommand } from "./lib/insert_memo_command.js";
 
 const main = async () => {
   const options = minimist(process.argv.slice(2));
@@ -17,7 +18,8 @@ const main = async () => {
     memoApp.addCommand(new DeleteMemoCommand());
     memoApp.command.execute();
   } else {
-    memoApp.insertMemo();
+    memoApp.addCommand(new InsertMemoCommand());
+    memoApp.command.execute();
   }
 };
